refactor: 파일 group 생성 및 캡슐화를 통한 코드 정리
