+++ BankManagerConsoleApp/BankManagerConsoleApp/BankManagerConsole/BankManagerConsole.swift
@@ -1,0 +1,13 @@
+//
+//  BankManagerConsole.swift
+//  BankManagerConsoleApp
+//
+//  Created by 수꿍, 브래드 on 2022/07/05.
+//
+
+struct BankManagerConsole {
+    mutating func run() {
+        showMenu()
+        chooseOption()
+    }
+}
