+++ BankManagerConsoleApp/BankManagerConsoleApp/Banker/Banker+Extension.swift
