+++ BankManagerConsoleApp/BankManagerConsoleApp/Banker/Banker.swift
@@ -1,5 +1,5 @@
 //
-//  Banker+Protocol.swift
+//  Banker.swift
 //  BankManagerConsoleApp
 //
 //  Created by 수꿍, 브래드 on 2022/07/01.
