+++ BankManagerConsoleApp/BankManagerConsoleApp/main.swift
@@ -3,6 +3,7 @@
 //  Created by 수꿍, 브래드. 
 //  Copyright © yagom academy. All rights reserved.
 // 
-var yagomBank = Bank(numberOfDepositBanker: 2, numberOfLoanBanker: 1)
 
-yagomBank.run()
+var bankManagerConsole = BankManagerConsole()
+
+bankManagerConsole.run()
